% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_measuredvalues_missingvalues.R
\name{barplot_samples_memi}
\alias{barplot_samples_memi}
\title{Barplot of number of measured/missing features of samples}
\usage{
barplot_samples_memi(tbl, measured = TRUE)
}
\arguments{
\item{tbl}{`tibble` object}

\item{measured}{`logical`, should the number of measured or missing values
be plotted}
}
\value{
`gg` object from `ggplot2`
}
\description{
`barplot_samples_memi` plots the number of measured/missing
features of samples as a barplot. The function will take as input the 
returned `tibble` of `samples_memi`.
}
\examples{
library(dplyr)
library(SummarizedExperiment)

## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
sample <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
featData <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment(assay = a, rowData = featData, colData = sample)

## create the data.frame with information on number of measured/missing
## values
tbl <- samples_memi(se) 

## plot number of measured values
barplot_samples_memi(tbl, measured = TRUE)

## plot number of missing values
barplot_samples_memi(tbl, measured = FALSE)
}
