% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{batchCorrectionAssay}
\alias{batchCorrectionAssay}
\title{Remove batch effects from (count/intensity) values of a 
\code{SummarizedExperiment}}
\usage{
batchCorrectionAssay(
  se,
  method = c("none", "removeBatchEffect (limma)"),
  batchColumn = colnames(se@colData)
)
}
\arguments{
\item{se}{\code{SummarizedExperiment}}

\item{method}{\code{character}, one of \code{"none"} or 
\code{"removeBatchEffect"}}

\item{batchColumn}{\code{character}, one of \code{colnames(colData(se))}}
}
\value{
\code{matrix}
}
\description{
The function \code{batchCorrectionAssay} removes the batch effect of 
(count/intensity) values of a \code{SummarizedExperiment}. 
It uses either the \code{removeBatchEffect} function 
or no batch effect correction method (pass-through, 
\code{none}).
}
\details{
The column \code{batchColumn} in \code{colData(se)} contains the information 
on the batch identity. Internal use in \code{shinyQC}.

If \code{batchColumn} is NULL, \code{batchColumn} is internally set to the 
name of the first column in \code{colData(se)} if 
\code{method = "removeBatchEffect (limma)"}.
}
\examples{
## create se
a <- matrix(seq_len(100), nrow = 10, ncol = 10, 
            dimnames = list(seq_len(10), paste("sample", seq_len(10))))
a[c(1, 5, 8), seq_len(5)] <- NA
set.seed(1)
a <- a + rnorm(100)
cD <- data.frame(name = colnames(a), 
    type = c(rep("1", 5), rep("2", 5)), batch = rep(c(1, 2), 5))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment::SummarizedExperiment(assay = a, 
    rowData = rD, colData = cD)

batchCorrectionAssay(se, method = "removeBatchEffect (limma)", 
                            batchColumn = "batch")

}
