% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{data}
\name{biocrates}
\alias{biocrates}
\alias{file}
\title{Convert Biocrates xlsx output to `SummarizedExperiment` object}
\format{
An object of class \code{character} of length 1.
}
\usage{
biocrates(file, sheetName = NULL, ...)
}
\arguments{
\item{file}{`character`}

\item{sheetName}{`character`, given to `read.xslx`}

\item{...}{additional parameters given to `read.xslx`}
}
\value{
`SummarizedExperiment` object
}
\description{
The function `biocrates` will create a `SummarizedExperiment` from a
Biocrates xlsx file. 
The function `biocrates` takes as input the path to a .xlsx file 
(Biocrates output) and additional parameters given to the `read.xlsx` 
function from the `xlsx` package (e.g. specifying the sheet name by
`sheetName` or the sheet index by `sheetIndex`).
}
\details{
The column "Sample Identification" has to contain unique identifiers
(no duplications).
}
\examples{
file <- "path/to/biocrates/object"

\dontrun{
biocrates(file = file, sheetName = NULL, ...)
}

}
\keyword{datasets}
