% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Create a boxplot of (count/intensity) values per sample}
\usage{
create_boxplot(x, title = "", log2 = TRUE, violin = FALSE)
}
\arguments{
\item{x}{`matrix` or `data.frame` containing the (count/intensity) values, 
samples are in columns and features in rows}

\item{title}{`character` or `numeric` of `length(1)`}

\item{log2}{`logical`, if `TRUE` (count/intensity) values are displayed as 
log2 values}

\item{violin}{`logical`, if `FALSE` a boxplot is created, if `TRUE` a 
violin plot is created}
}
\value{
`gg`
}
\description{
The function `create_boxplot` creates
}
\details{
Internal usage in `shinyQC`.
}
\examples{
x <- matrix(1:100, ncol = 10, dimnames = list(1:10, paste("sample", 1:10)))
create_boxplot(x, title = "", log2 = TRUE, violin = FALSE)

}
