% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{cv}
\alias{cv}
\title{Calculate coefficient of variation}
\usage{
cv(x, name = "raw")
}
\arguments{
\item{x}{`matrix`}

\item{name}{`character`, the name of the returned list}
}
\value{
`list`
}
\description{
The function `cv` calculates the coefficient of variation from columns of 
a matrix. The coefficients of variation are calculated according to the 
formula `sd(y) / mean(y) * 100` with `y` the column values.
}
\details{
The function returned a named `list` (the name is specified by the `name`
argument) containing the coefficient of variation of the columns of `x`.
}
\examples{
x <- matrix(1:10, ncol = 2)
cv(x)

}
