% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{ECDF}
\alias{ECDF}
\title{Create ECDF plot of a sample against a reference}
\usage{
ECDF(se, sample = colnames(se), group = c("all", colnames(colData(se))))
}
\arguments{
\item{se}{`SummarizedExperiment` object}

\item{sample}{`character`, name of the sample to compare against the group}

\item{group}{`character`, either `"all"` or one of `colnames(colData(se))`}
}
\value{
`gg` object from `ggplot2`
}
\description{
The function `ECDF` creates a plot of the empirical cumulative distribution
function of a specified sample and an outgroup.
}
\details{
Internal use in `shinyQC`.
}
\examples{
## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
    dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
sample <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
featData <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment(assay = a, rowData = featData, colData = sample)

ECDF(se, "sample 1", group = "all")

}
