% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_measuredvalues_missingvalues.R
\name{hist_feature_category}
\alias{hist_feature_category}
\title{Histogram of features per sample type}
\usage{
hist_feature_category(se, measured = TRUE, category = "type", ...)
}
\arguments{
\item{se}{`SummarizedExperiment`, the assay slot contains the intensity 
values per sample. Missing values are encoded as `NA`.}

\item{measured}{`logical`, should the measured values 
(`measured = TRUE`) or missing values (`measured = FALSE`) be taken}

\item{category}{`character`, corresponding to a column in `colData(se)`}

\item{...}{additional parameters passed to `geom_histogram`, e.g. `binwidth`.}
}
\value{
`gg` object from `ggplot2`
}
\description{
The function `hist_feature_category` creates histogram
plots for each sample type in `se`.
}
\examples{

## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
cD <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment::SummarizedExperiment(assay = a, 
    rowData = rD, colData = cD)

hist_feature_category(se, measured = TRUE, category = "type")

}
