% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_samplesfeatures.R
\name{hist_sample}
\alias{hist_sample}
\alias{hist_sample_num}
\title{Return the number of a category}
\usage{
hist_sample_num(se, category = "type")

hist_sample(tbl, category = "type")
}
\arguments{
\item{se}{`SummarizedExperiment` object}

\item{category}{`character`, x-axis label of the plot}

\item{tbl}{`tibble` as returned by `hist_sample_num`}
}
\value{
`tibble`

`gg` object from `ggplot2`
}
\description{
`hist_sample_num` returns the number of a category 
(e.g. sample types) as a `tibble`.
The function will retrieve first a column in `colData(se)`. 
The function will return a `tibble` containing the numerical 
values of the quantities.

`hist_sample` plots the number of a category (e.g. sample types)
as a histogram. It use the returned `tibble` from `hist_sample_num`.
}
\examples{
## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
cD <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment::SummarizedExperiment(assay = a, 
    rowData = rD, colData = cD)

hist_sample_num(se, category = "type")

## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
cD <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment::SummarizedExperiment(assay = a, 
    rowData = rD, colData = cD)

tbl <- hist_sample_num(se, category = "type")
hist_sample(tbl)

}
