% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{impute}
\alias{impute}
\title{Impute missing values in a `data.frame`, `tibble`, `matrix`}
\usage{
impute(x, method = c("BPCA", "kNN", "MLE", "Min", "MinDet", "MinProb"))
}
\arguments{
\item{x}{`data.frame`, `tibble`, or `matrix`}

\item{method}{`character`, one of `"BPCA"`, `"kNN"`, `"MLE`, `"Min"`, 
`"MinDet"`, or `"MinProb"`}
}
\value{
`matrix`
}
\description{
The function `impute` imputes missing values based on one of the following 
principles: Bayesian missing value imputation (`BPCA`), k-nearest 
neighbor averaging (`KNN`), Malimum likelihood-based imputation method using
the EM algorithm (`MLE`), replacement by the smallest non-missing value
in the data (`Min`), replacement by the minimal value observed as
the q-th quantile (`MinDet`, default `q = 0.01`), and replacement by
random draws from a Gaussian distribution centred to a minimal value 
(`MinProb`). 

The function is a wrapper function from the functions `impute_bpca`, 
`impute_knn`, `impute_mle`, and `impute_min` from the `MsCoreUtils` package
and the functions `impute.MinDet` and `impute.MinProb` from the `imputeLCMD`
package.
}
\examples{
x <- matrix(1:100, nrow = 10, ncol = 10, 
    dimnames = list(1:10, paste("sample", 1:10)))
x[c(1, 5, 8), 1:5] <- NA
impute(x, method = "kNN")

}
