% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{MAvalues}
\alias{MAvalues}
\title{Create values (M and A) for MA plot}
\usage{
MAvalues(se, log2 = TRUE, group = c("all", colnames(colData(se))))
}
\arguments{
\item{se}{`SummarizedExperiment`}

\item{log2}{`logical`}

\item{group}{`character`, either `"all"` or one of `colnames(colData(se))`}
}
\value{
`tibble`
}
\description{
THe function `MAvalues` will create MA values as input for the function 
`MAplot`. `M` and `A` are specified relative to specified samples which 
is determined by the `group` argument. In case of `group == "all"`, all 
samples (expect the specified one) are taken for the reference calculation.
In case of `group != "all"` will use the samples belonging to the same group
given in `colnames(colData(se))` expect the specified one.
}
\examples{
## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
sample <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
featData <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment(assay = a, rowData = featData, colData = sample)

MAvalues(se, group = "all")

}
