% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{maxQuant}
\alias{maxQuant}
\title{Convert MaxQuant xlsx output to \code{SummarizedExperiment} object}
\usage{
maxQuant(
  file,
  intensity = c("iBAQ", "LFQ"),
  sheet,
  type = c("txt", "xlsx"),
  ...
)
}
\arguments{
\item{file}{\code{character}}

\item{intensity}{\code{character}, either \code{"iBAQ"} or \code{"LFQ"}}

\item{sheet}{\code{character} or \code{numeric}, the name or index of the 
sheet to read data from}

\item{type}{\code{character}, either \code{"txt"} or \code{"xlsx"}}

\item{...}{additional parameters given to \code{read.xlsx} (for 
\code{type = "xlsx"})}
}
\value{
\code{SummarizedExperiment} object
}
\description{
The function \code{maxQuant} will create a \code{SummarizedExperiment} from a
MaxQuant xlsx file. 
The function \code{maxQuant} takes as input the path to a .xlsx file 
(MaxQuant output) and additional parameters given to the \code{read.xlsx} 
function from the \code{openxlsx} package (e.g. specifying the sheet name 
or index by \code{sheet}).
}
\details{
The argument \code{intensity} will specify if the \code{iBAQ} or 
\code{LFQ} values are taken. 

The argument \code{type} will specify if the data is loaded from \code{txt} 
or \code{xlsx} files.
}
\examples{
file <- "path/to/maxQuant/object.txt"
\donttest{maxQuant(file = file, intensity = "iBAQ", type = "txt")}

}
