% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_measuredvalues_missingvalues.R
\name{measured_category}
\alias{measured_category}
\title{Obtain the number of measured intensities per sample type}
\usage{
measured_category(se, measured = TRUE, category = "type")
}
\arguments{
\item{se}{`SummarizedExperiment`}

\item{measured}{`logical`, should the measured values 
(`measured = TRUE`) or missing values (`measured = FALSE`) be taken}

\item{category}{`character`, corresponds to a column name in `colData(se)`}
}
\value{
`tibble`
}
\description{
The function `measured_category` creates a `tibble` with
the number of measured values per feature. 0 means that there were only 
missing values (`NA`) for the feature and sample type. 
`measured_category` will return a `tibble` where columns are the 
unique sample types and rows are the features as in `assay(se)`.
}
\details{
`measured_category` is a helper function.
}
\examples{
library(dplyr)
library(SummarizedExperiment)

## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
sample <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
featData <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment(assay = a, rowData = featData, colData = sample)

measured_category(se, measured = TRUE, category = "type")

}
