% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{normalize}
\alias{normalize}
\title{Normalize a data sets (reduce technical sample effects)}
\usage{
normalize(x, method = c("none", "sum", "quantile division", "quantile"), probs)
}
\arguments{
\item{x}{`data.frame`, `tibble`, or `matrix` with samples in columns and 
features in rows}

\item{method}{`character`, one of `"none"`, `"sum"`, `"quantile division"`, 
`"quantile"`}

\item{probs}{`numeric`, ranging between `[0, 1)`. `probs` is used as the 
divisor for quantile division in `method = "quantile division"`}
}
\value{
`matrix`
}
\description{
The function `normalize` performs normalization by sum of the 
(count/intensity) values per sample or quantile division per sample
or by quantile normalization (adjusting the distributions that they become
identical in statistical distributions). The divisor for quantile division
(e.g., the 75% quantile per sample) can be specified by the `probs` argument.
Quantile normalization is performed by using the `normalizeQuantiles` 
function from `limma`.
}
\details{
Internal usage in `shinyQC`. If `method` is set to `"none"`, the object
`x` is returned as is (pass-through).
}
\examples{
x <- matrix(1:100, nrow = 10, ncol = 10, 
        dimnames = list(1:10, paste("sample", 1:10)))
normalize(x, "sum")

}
