% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_dimensionreduction.R
\name{ordination}
\alias{ordination}
\title{Dimensionality reduction with ordination methods PCA,
PCoA, NMDS, UMAP and tSNE}
\usage{
ordination(x, type = c("PCA", "PCoA", "NMDS", "tSNE", "UMAP"), params = list())
}
\arguments{
\item{x}{\code{matrix}, containing no missing values, samples are in columns 
and features are in rows}

\item{type}{\code{character}, specifying the type/method to use for 
dimensionality reduction. One of \code{PCA}, \code{PCoA}, \code{NMDS}, 
\code{tSNE}, or \code{UMAP}.}

\item{params}{\code{list}, arguments/parameters given to the functions 
\code{stats::prcomp}, \code{stats::dist}, \code{Rtsne::Rtsne}, 
\code{umap::umap}}
}
\value{
\code{tbl}
}
\description{
The function \code{ordination} creates a \code{data.frame} 
with the coordinates of the projected data. The function allows for the 
following projections:
Principal Component Analysis (PCA), Principal Coordinates 
Analysis/Multidimensional Scaling (PCoA), Non-metric Multidimensional 
scaling (NMDS), t-distributed stochastic neighbor embedding (tSNE), and 
Uniform Manifold Approimation and Projection (UMAP).
}
\details{
The function \code{ordination} is a wrapper around the following 
functions \code{stats::prcomp} (PCA), \code{stats::cmdscale} (PCoA), 
\code{vegan::metaMDS} (NMDS), \code{Rtsne::Rtsne} (tSNE), and 
\code{umap::umap} (UMAP). For the function \code{umap::umap} 
the method is set to \code{naive}.
}
\examples{
x <- matrix(rnorm(1:10000), ncol = 100)
rownames(x) <- paste("feature", 1:nrow(x))
colnames(x) <- paste("sample", 1:ncol(x))
params <- list(method = "euclidean", ## dist
    initial_dims = 10, max_iter = 100, dims = 3, perplexity = 3, ## tSNE
    min_dist = 0.1, n_neighbors = 15, spread = 1) ## UMAP
ordination(x, type = "PCA", params = params)
ordination(x, type = "PCoA", params = params)
ordination(x, type = "NMDS", params = params)
ordination(x, type = "tSNE", params = params)
ordination(x, type = "UMAP", params = params)

}
\author{
Thomas Naake
}
