% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{spectronaut}
\alias{spectronaut}
\title{Convert Spectronaut xlsx output to \code{SummarizedExperiment} object}
\usage{
spectronaut(file, sheetIntensities, sheetAnnotation, ...)
}
\arguments{
\item{file}{\code{character}}

\item{sheetIntensities}{\code{character} or \code{numeric}, name or index of the 
sheet where the intensities are stored}

\item{sheetAnnotation}{\code{character} or \code{numeric}, name or index of the 
sheet where the annotations are stored}

\item{...}{additional parameters given to \code{read.xslx}}
}
\value{
\code{SummarizedExperiment} object
}
\description{
The function \code{spectronaut} will create a \code{SummarizedExperiment} 
from a Spectronaut xlsx file. 
The function \code{spectronaut} takes as input the path to a .xlsx file 
(Spectronaut output).
}
\details{
The function requires that the intensity values are stored in the sheet
\code{sheetIntensities} and the sample annotations in the sheet
\code{sheetAnnotation}.

The sample names are taken from the column \code{"SAMPLE_IDs"} from the 
sheet \code{sheetAnnotation}.
}
\examples{
file <- "path/to/spectronaut/object"
\donttest{spectronaut(file = file, sheetIntensitities = 1, 
    sheetAnnotation = 2, ...)}

}
