% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_values.R
\name{tP_meanSdUI}
\alias{tP_meanSdUI}
\title{Tab panel UI for tab panel 'mean-sd plot'}
\usage{
tP_meanSdUI(id)
}
\value{
`shiny.tag` with HTML content
}
\description{
The module defines the UI in the tab panel 'mean-sd plot'. It serves as a 
wrapper for the function `box_meanSdUI`.
}
\details{
`tP_meanSdUI` returns the HTML code for the tab-pane 'mean-sd plot'. 
Internal function for `shinyQC`.
}
\examples{
tP_meanSdUI()

}
\author{
Thomas Naake
}
