% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{transform}
\alias{transform}
\title{Transform the (count/intensity) values of a `data.frame`, `tbl` or
`matrix`}
\usage{
transform(x, method = c("none", "log2", "vsn"))
}
\arguments{
\item{x}{`data.frame`, `tibble`, or `matrix` with samples in columns and 
features in rows}

\item{method}{`character`, one of `"none"`, `"log2"` or `"vsn"`}
}
\value{
`matrix`
}
\description{
The function `transform` transforms the (count/intensity) values of a 
`data.frame`, `tbl` or `matrix`. It uses either `log2`, variance 
stabilizing normalisation (`vsn`) or no transformation method (pass-through,
`none`). The object
`x` has the samples in the columns and the features in the rows.
}
\details{
Internal use in `shinyQC`.
}
\examples{
x <- matrix(1:1000, nrow = 100, ncol = 10, 
        dimnames = list(1:100, paste("sample", 1:10)))
transform(x, "none")
transform(x, "log2")
transform(x, "vsn")

}
