% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_measuredvalues_missingvalues.R
\name{upset_category}
\alias{upset_category}
\title{UpSet plot to display measures values across sample types}
\usage{
upset_category(se, category = "type", ...)
}
\arguments{
\item{se}{`SummarizedExperiment`,
 containing the intensity values in `assay(se)`, missing values are
encoded by `NA`}

\item{category}{`character`, corresponding to a column in `colData(se)`}

\item{...}{additional parameters passed to `measured_category`}
}
\value{
`UpSet` plot
}
\description{
The function `upset_category` displays the frequency of measured values per
feature with respect to class/sample type to assess difference in
occurences. Internally, the measured values per sample are obtained via
the `measured_category` function: this function will
}
\examples{
## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
cD <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment::SummarizedExperiment(assay = a, 
    rowData = rD, colData = cD)

upset_category(se, category = "type")

}
