% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{batch}
\alias{batch}
\title{Remove batch effects from (count/intensity) values of a 
`SummarizedExperiment`}
\usage{
batch(se, method = c("none", "removeBatchEffect (limma)"), batchColumn)
}
\arguments{
\item{se}{`SummarizedExperiment`, `tibble`, or `matrix`}

\item{method}{`character`, one of `"none"` or `"removeBatchEffect"`}

\item{batchColumn}{`character`, one of}
}
\value{
`matrix`
}
\description{
The function `batch` removes the batch effect of (count/intensity) values of 
a `SummarizedExperiment`. It uses either the `removeBatchEffect` function 
or no batch effect correction method (pass-through, 
`none`).
}
\details{
Internal use in `shinyQC`.
}
\examples{
x <- matrix(1:1000, nrow = 100, ncol = 10, 
        dimnames = list(1:100, paste("sample", 1:10)))
transform(x, "none")
transform(x, "log2")
transform(x, "vsn")

}
