% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{biocrates}
\alias{biocrates}
\title{Convert Biocrates xlsx output to `SummarizedExperiment` object}
\usage{
biocrates(file, sheet, ...)
}
\arguments{
\item{file}{`character`}

\item{sheet}{`character` or `numeric`, the name or index of the sheet to 
read data from}

\item{...}{additional parameters given to `read.xslx`}
}
\value{
`SummarizedExperiment` object
}
\description{
The function `biocrates` will create a `SummarizedExperiment` from a
Biocrates xlsx file. 
The function `biocrates` takes as input the path to a .xlsx file 
(Biocrates output) and additional parameters given to the `read.xlsx` 
function from the `openxlsx` package (e.g. specifying the sheet name or index
by `sheet`).
}
\details{
The column "Sample Identification" has to contain unique identifiers
(no duplications).
}
\examples{
file <- "path/to/biocrates/object"
\donttest{biocrates(file = file, sheet = 1)}

}
