% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{biocrates}
\alias{biocrates}
\title{Convert Biocrates xlsx output to \code{SummarizedExperiment} object}
\usage{
biocrates(file, sheet, ...)
}
\arguments{
\item{file}{\code{character}}

\item{sheet}{\code{character} or \code{numeric}, the name or index of the 
sheet to read data from}

\item{...}{additional parameters given to \code{read.xlsx}}
}
\value{
\code{SummarizedExperiment} object
}
\description{
The function \code{biocrates} will create a \code{SummarizedExperiment} 
from a Biocrates xlsx file. 
The function \code{biocrates} takes as input the path to a .xlsx file 
(Biocrates output) and additional parameters given to the \code{read.xlsx} 
function from the \code{openxlsx} package (e.g. specifying the sheet name or 
index by \code{sheet}).
}
\details{
The column "Sample Identification" has to contain unique identifiers
(no duplications).
}
\examples{
file <- "path/to/biocrates/object"
\donttest{biocrates(file = file, sheet = 1)}

}
