% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{createBoxplot}
\alias{createBoxplot}
\title{Create a boxplot of (count/intensity) values per sample}
\usage{
createBoxplot(
  se,
  orderCategory = colnames(colData(se)),
  title = "",
  log2 = TRUE,
  violin = FALSE
)
}
\arguments{
\item{se}{`SummarizedExperiment` containing the (count/intensity) values in
the `assay` slot}

\item{orderCategory}{`character`, one of `colnames(colData(se))`}

\item{title}{`character` or `numeric` of `length(1)`}

\item{log2}{`logical`, if `TRUE` (count/intensity) values are displayed as 
log2 values}

\item{violin}{`logical`, if `FALSE` a boxplot is created, if `TRUE` a 
violin plot is created}
}
\value{
`gg` object from `ggplot2`
}
\description{
The function `create_boxplot` creates a boxplot per sample for the 
intensity/count values.
}
\details{
Internal usage in `shinyQC`.
}
\examples{
## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
    dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
cD <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment::SummarizedExperiment(assay = a, 
    rowData = rD, colData = cD)

createBoxplot(se, orderCategory = "name", title = "", log2 = TRUE, 
    violin = FALSE)

}
