% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{distShiny}
\alias{distShiny}
\title{Create distance matrix from numerical matrix}
\usage{
distShiny(x, method = "euclidean")
}
\arguments{
\item{x}{`matrix` or `data.frame`}

\item{method}{`character`, method for distance calculation}
}
\description{
The function `distShiny` takes as an input a numerical `matrix` or
`data.frame` and returns the distances between the rows and columns based
on the defined `method` (e.g. euclidean distance).
}
\details{
Internal use in `shinyQC`.
}
\examples{
x <- matrix(1:100, nrow = 10, ncol = 10, 
        dimnames = list(1:10, paste("sample", 1:10)))
distShiny(x = x)

}
