% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{ECDF}
\alias{ECDF}
\title{Create ECDF plot of a sample against a reference}
\usage{
ECDF(se, sample = colnames(se), group = c("all", colnames(colData(se))))
}
\arguments{
\item{se}{`SummarizedExperiment` object}

\item{sample}{`character`, name of the sample to compare against the group}

\item{group}{`character`, either `"all"` or one of `colnames(colData(se))`}
}
\value{
`gg` object from `ggplot2`
}
\description{
The function `ECDF` creates a plot of the empirical cumulative distribution
function of a specified sample and an outgroup (reference). The reference
is specified by the `group` argument. The row-wise (feature) mean values of
the reference are calculated after excluding the specified `sample`.
}
\details{
Internal use in `shinyQC`.
}
\examples{
## create se
set.seed(1)
a <- matrix(rnorm(1000), nrow = 100, ncol = 10, 
    dimnames = list(1:100, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
cD <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment(assay = a, rowData = rD, colData = cD)

ECDF(se, sample = "sample 1", group = "all")

}
