% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_dimensionreduction.R
\name{explVar}
\alias{explVar}
\title{Retrieve the explained variance for each principal component}
\usage{
explVar(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{`matrix`, containing no missing values (`NA`)}

\item{center}{`logical`, indicating whether the variables should be shifted 
to be zero centered}

\item{scale}{`logical`, indicating whether the variables should be shifted
to be zero centered}
}
\value{
`numeric` vector with the proportion of explained variance for each 
PC
}
\description{
The function `explVar` calculates the proportion of explained variance for
each principal component (PC).
}
\details{
`explVar` uses the function `prcomp` from the `stats` package to retrieve
the explained standard deviation per PC.
}
\examples{
x <- matrix(1:100, nrow = 10, ncol = 10, 
    dimnames = list(1:10, paste("sample", 1:10)))
set.seed(1)
x <- x + rnorm(100)
explVar(x = x, center = TRUE, scale = TRUE)

}
