% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_measuredvalues_missingvalues.R
\name{histFeature}
\alias{histFeature}
\title{Histogram for measured value per feature}
\usage{
histFeature(x, measured = TRUE, ...)
}
\arguments{
\item{x}{`matrix` containing intensities. Missing values are encoded 
as `NA`.}

\item{measured}{`logical`, should the measured values 
(`measured = TRUE`) or missing values (`measured = FALSE`) be taken}

\item{...}{additional parameters passed to `geom_histogram`, e.g. 
`binwidth`.}
}
\value{
`plotly` object from `ggplotly`
}
\description{
The function `histFeature` creates a histogram with the number
of measured/missing values per feature.
}
\examples{
x <- matrix(c(c(1, 1, 1), c(1, NA, 1), c(1, NA, 1), 
    c(1, 1, 1), c(NA, 1, 1), c(NA, 1, 1)), byrow = FALSE, nrow = 3)
colnames(x) <- c("A_1", "A_2", "A_3", "B_1", "B_2", "B_3")
histFeature(x, binwidth = 1)

}
