% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{hoeffDValues}
\alias{hoeffDValues}
\title{Create values of Hoeffding's D statistics from M and A values}
\usage{
hoeffDValues(tbl, name = "raw")
}
\arguments{
\item{tbl}{`tibble`, as obtained from the function `MAvalues`}

\item{name}{`character`, name of the returned lists}
}
\value{
named list of lists
}
\description{
The function creates and returns Hoeffding's D statistics values 
from MA values.
}
\details{
The function uses the function `hoeffd` from the `Hmisc` package to 
calculate the values.
}
\examples{
library(dplyr)
library(SummarizedExperiment)

## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
sample <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
featData <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment(assay = a, rowData = featData, colData = sample)

tbl <- MAvalues(se)
hd_r <- hoeffDValues(tbl, "raw")

## normalized values
se_n <- se
assay(se_n) <- assay(se) \%>\% normalize(., "sum")
tbl_n <- MAvalues(se_n, group = "all")
hd_n <- hoeffDValues(tbl_n, "normalized")

## transformed values
se_t <- se
assay(se_t) <- assay(se_n) \%>\% transform(., "log2")
tbl_t <- MAvalues(se_t, group = "all")
hd_t <- hoeffDValues(tbl_t, "transformed")

}
