% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{maxQuant}
\alias{maxQuant}
\title{Convert MaxQuant xlsx output to `SummarizedExperiment` object}
\usage{
maxQuant(file, type = c("iBAQ", "LFQ"), sheetName = NULL, ...)
}
\arguments{
\item{file}{`character`}

\item{type}{`character`, either `iBAQ` or `LFQ`}

\item{sheetName}{`character`, given to `read.xslx`}

\item{...}{additional parameters given to `read.xslx`}
}
\value{
`SummarizedExperiment` object
}
\description{
The function `maxQuant` will create a `SummarizedExperiment` from a
MaxQuant xlsx file. 
The function `maxQuant` takes as input the path to a .xlsx file 
(MaxQuant output) and additional parameters given to the `read.xlsx` 
function from the `xlsx` package (e.g. specifying the sheet name by
`sheetName` or the sheet index by `sheetIndex`).
}
\details{
The argument `type` will specify if the `iBAQ` or `LFQ` values are taken.
}
\examples{
file <- "path/to/maxQuant/object"
\donttest{maxQuant(file = file, sheetName = NULL)}

}
