% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{maxQuant}
\alias{maxQuant}
\title{Convert MaxQuant xlsx output to `SummarizedExperiment` object}
\usage{
maxQuant(file, type = c("iBAQ", "LFQ"), sheet, ...)
}
\arguments{
\item{file}{`character`}

\item{intensity}{`character`, either `"iBAQ"` or `"LFQ"`}

\item{sheet}{`character` or `numeric`, the name or index of the sheet to 
read data from}

\item{type}{`character`, either `"txt"` or `"xlsx"`}

\item{...}{additional parameters given to `read.xslx` (for `type = "xlsx"`)}
}
\value{
`SummarizedExperiment` object
}
\description{
The function `maxQuant` will create a `SummarizedExperiment` from a
MaxQuant xlsx file. 
The function `maxQuant` takes as input the path to a .xlsx file 
(MaxQuant output) and additional parameters given to the `read.xlsx` 
function from the `openxlsx` package (e.g. specifying the sheet name or index 
by `sheet`).
}
\details{
The argument `intensity` will specify if the `iBAQ` or `LFQ` values are taken. 

The argument `type` will specify if the data is loaded from `txt` or `xlsx`
files.
}
\examples{
file <- "path/to/maxQuant/object.txt"
\donttest{maxQuant(file = file, type = "iBAQ", sheet = 1)}

}
