% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_samplesfeatures.R
\name{mosaic}
\alias{mosaic}
\title{Mosaic plot for two factors in colData(se)}
\usage{
mosaic(se, f1, f2)
}
\arguments{
\item{se}{`SummarizedExperiment` object}

\item{f1}{`character`, `f1` is one of the column names in `colData(se)`}

\item{f2}{`character`, `f2` is one of the column names in `colData(se)`}
}
\value{
`gg` object from `ggplot2`
}
\description{
The function `mosaic` creates a mosaic plot of two factors from
an `SummararizedExperiment` object. The columns `f1` and `f2` 
are taken from `colData(se)`.
}
\details{
Code partly taken from
https://stackoverflow.com/questions/21588096/pass-string-to-facet-grid-ggplot2
}
\examples{
## create se
a <- matrix(1:100, nrow = 10, ncol = 10, 
            dimnames = list(1:10, paste("sample", 1:10)))
a[c(1, 5, 8), 1:5] <- NA
set.seed(1)
a <- a + rnorm(100)
cD <- data.frame(name = colnames(a), 
    type = c(rep("1", 5), rep("2", 5)),
    cell_type = c("A", "B"))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment::SummarizedExperiment(assay = a, 
    rowData = rD, colData = cD)

mosaic(se, "cell_type", "type")

}
