% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_values.R
\name{normalizeAssay}
\alias{normalizeAssay}
\title{Normalize a data sets (reduce technical sample effects)}
\usage{
normalizeAssay(
  a,
  method = c("none", "sum", "quantile division", "quantile"),
  probs = 0.75
)
}
\arguments{
\item{a}{\code{matrix} with samples in columns and features in rows}

\item{method}{\code{character}, one of \code{"none"}, \code{"sum"}, 
\code{"quantile division"}, \code{"quantile"}}

\item{probs}{\code{numeric}, ranging between \code{[0, 1)}. \code{probs} 
is used as the divisor for quantile division in 
\code{method = "quantile division"}}
}
\value{
\code{matrix}
}
\description{
The function \code{normalizeAssay} performs normalization by sum of the 
(count/intensity) values per sample or quantile division per sample
or by quantile normalization (adjusting the distributions that they become
identical in statistical distributions). The divisor for quantile division
(e.g., the 75% quantile per sample) can be specified by the \code{probs} 
argument. Quantile normalization is performed by using the 
\code{normalizeQuantiles} function from \code{limma}.
}
\details{
Internal usage in \code{shinyQC}. If \code{method} is set to \code{"none"}, 
the object \code{x} is returned as is (pass-through).

If \code{probs} is NULL, \code{probs} is internally set to "name" if 
\code{method = "quantile division"}.
}
\examples{
a <- matrix(1:100, nrow = 10, ncol = 10, 
        dimnames = list(1:10, paste("sample", 1:10)))
normalizeAssay(a, "sum")

}
