% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_dimensionreduction.R
\name{ordinationPlot}
\alias{ordinationPlot}
\title{Plot the coordinates from `ordination` values}
\usage{
ordinationPlot(
  tbl,
  se,
  highlight = c("none", colnames(colData(se))),
  explainedVar = NULL,
  x_coord,
  y_coord,
  height = 600
)
}
\arguments{
\item{tbl}{`tbl` as obtained by the function `ordination`}

\item{se}{`SummarizedExperiment`}

\item{highlight}{`character`, one of `"none"` or `colnames(colData(se))`}

\item{explainedVar}{NULL or named `numeric`, if `numeric` `explainedVar` 
contains the explained variance per principal component (names of 
`explainedVar` corresponds to the principal components)}

\item{x_coord}{`character`, column name of `tbl` that stores x coordinates}

\item{y_coord}{`character`, column name of `tbl` that stores y coordinates}

\item{height}{`numeric`, specifying the height of the plot (in pixels)}
}
\value{
`plotly`
}
\description{
The function `ordinationPlot` creates a dimension reduction plot. 
The function takes as input the `tbl` object obtained 
from the `ordination` function. The `tbl` contains transformed
values by one of the ordination methods.
}
\details{
The function `ordinationPlot` is a wrapper for a `ggplot`/`ggplotly` 
expression.
}
\examples{
library(SummarizedExperiment)

## create se
a <- matrix(1:100, nrow = 10, ncol = 10, byrow = TRUE,
            dimnames = list(1:10, paste("sample", 1:10)))
set.seed(1)
a <- a + rnorm(100)
cD <- data.frame(name = colnames(a), type = c(rep("1", 5), rep("2", 5)))
rD <- data.frame(spectra = rownames(a))
se <- SummarizedExperiment(assay = a, rowData = rD, colData = cD)

pca <- ordination(x = assay(se), type = "PCA", params = list())

ordinationPlot(tbl = pca, se = se, highlight = "type", 
    x_coord = "PC1", y_coord = "PC2")

}
\author{
Thomas Naake
}
