% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_dimensionreduction.R
\name{permuteExplVar}
\alias{permuteExplVar}
\title{Permute the expression values and retrieve the explained variance}
\usage{
permuteExplVar(x, n = 10, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{`matrix` or `data.frame`}

\item{n}{`numeric`, number of permutation rounds}

\item{center}{`logical`, passed to the function `explVar`}

\item{scale}{`logical`, passed to the function `explVar`}
}
\value{
matrix with explained variance
}
\description{
The function `permuteExplVar` determines the explained variance of the 
permuted expression matrix (`x`). It is used to determine the optimal 
number of PCs for tSNE.
}
\details{
For the input of tSNE, typically, we want to reduce the 
initial number of dimensions linearly with PCA (used as the `initial_dims`
arguments in the `Rtsne` funtion). The reduced data set is used for feeding 
into tSNE. By plotting the percentage of variance explained by the Princical
Components (PCs) we can estimate how many PCs we keep as input into tSNE.
However, if we select too many PCs, noise will be included as input to tSNE;
if we select too few PCs we might loose the important data structures. 
To get a better understanding how many PCs to include, randomization will
be employed and the observed variance will be compared to the permuted 
variance.
}
\examples{
 
x <- matrix(1:100, nrow = 10, ncol = 10, 
    dimnames = list(1:10, paste("sample", 1:10)))
permuteExplVar(x = x, n = 10, center = TRUE, scale = TRUE)

}
