% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_dimensionreduction.R
\name{plotPCAVar}
\alias{plotPCAVar}
\title{Plot of explained variance against the principal components}
\usage{
plotPCAVar(var_x, var_perm = NULL)
}
\arguments{
\item{var_x}{`numeric` (named `numeric` vector)}

\item{var_perm}{`matrix` with the explained variance obtained by permutation
(function `permuteExplVar`)}
}
\value{
`gg` object from `ggplot`
}
\description{
The function `plotPCAVar` plots the explained variance (in %) on the y-axis
against the principal components for the measured and permuted values.
}
\details{
The argument `var_perm` is optional and visualization of permuted values 
can be omitted by setting `var_perm = NULL`.
}
\examples{
x <- matrix(1:100, ncol = 10)
var_x <- explVar(x = x, center = TRUE, scale = TRUE)
var_perm <- permuteExplVar(x = x, n = 100, center = TRUE, scale = TRUE)
plotPCAVar(var_x = var_x, var_perm = var_perm)

}
