% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{spectronaut}
\alias{spectronaut}
\title{Convert Spectronaut xlsx output to `SummarizedExperiment` object}
\usage{
spectronaut(file, sheetIntensities, sheetAnnotation, ...)
}
\arguments{
\item{file}{`character`}

\item{sheetIntensities}{`character` or `numeric`, name or index of the 
sheet where the intensities are stored}

\item{sheetAnnotation}{`character` or `numeric`, name or index of the 
sheet where the annotations are stored}

\item{...}{additional parameters given to `read.xslx`}
}
\value{
`SummarizedExperiment` object
}
\description{
The function `spectronaut` will create a `SummarizedExperiment` from a
Spectronaut xlsx file. 
The function `spectronaut` takes as input the path to a .xlsx file 
(Spectronaut output).
}
\details{
The function requires that the intensity values are stored in the sheet
`sheetIntensities` and the sample annotations in the sheet
`sheetAnnotation`.

The sample names are taken from the column `"SAMPLE_IDs"` from the 
sheet `sheetAnnotation`.
}
\examples{
file <- "path/to/spectronaut/object"
\donttest{spectronaut(file = file, sheetIntensitities = 1, 
    sheetAnnotation = 2, ...)}

}
